Fix pickle error with HPOneViewException
This commit fixes #320.

The problem seems to be that when oneview_response is not None the
Exception base class is called with two arguments whereas
HPOneViewException is defined to have one argument.  This fix overrides
the __reduce__ method so that pickle will recognise the situation where
two arguments are used.