Allow connection reuse
We have to use the SDK in environments where the total number of
connections is restricted, where repeatedly opening and closing
connections can cause instability, and with fairly large latencies in
opening a new connection.  To aid stability and SDK performance in these
environments we allow for the reuse of https connections as returned by
get_connection().  Connections will be reused when 'reuse_connection' is
set to True or when the ONEVIEWSDK_REUSE_CONNECTION is anything other
than an empty string.  We've added a unit-test for this new capability.